+++ app/components/Views/AccountPermissions/AccountPermissionsConnected/AccountPermissionsConnected.tsx
@@ -7,7 +7,6 @@ import SheetActions from '../../../../component-library/components-temp/SheetAct
 import { strings } from '../../../../../locales/i18n';
 import AccountSelectorList from '../../../../components/UI/AccountSelectorList';
 import { AccountPermissionsScreens } from '../AccountPermissions.types';
-import { switchActiveAccounts } from '../../../../core/Permissions';
 import {
   ToastContext,
   ToastVariants,
@@ -30,6 +29,7 @@ import Button, {
   ButtonVariants,
   ButtonWidthTypes,
 } from '../../../../component-library/components/Buttons/Button';
+import Engine from '../../../../core/Engine';
 
 const AccountPermissionsConnected = ({
   ensByAccountAddress,
@@ -54,7 +54,7 @@ const AccountPermissionsConnected = ({
   const switchActiveAccount = useCallback(
     (address: string) => {
       if (address !== activeAddress) {
-        switchActiveAccounts(hostname, address);
+        Engine.setSelectedAddress(address);
       }
       onDismissSheet();
       const activeAccountName = getAccountNameWithENS({
@@ -81,7 +81,6 @@ const AccountPermissionsConnected = ({
       onDismissSheet,
       accounts,
       ensByAccountAddress,
-      hostname,
       toastRef,
       accountAvatarType,
     ],