+++ app/core/RPCMethods/wallet_watchAsset.test.ts
@@ -1,6 +1,6 @@
 import { RpcEndpointType } from '@metamask/network-controller';
 import Engine from '../Engine';
-import wallet_watchAsset from './wallet_watchAsset';
+import { wallet_watchAsset } from './wallet_watchAsset';
 // eslint-disable-next-line import/no-namespace
 import * as transactionsUtils from '../../util/transactions';
 import {