+++ app/util/test/keyringControllerTestUtils.ts
@@ -60,6 +60,8 @@ const MOCK_DEFAULT_KEYRINGS_METADATA: KeyringMetadata[] = [
 ];
 
 export const MOCK_KEYRING_CONTROLLER_STATE = {
+  isUnlocked: jest.fn(),
+  getAccountKeyringType: jest.fn(),
   keyring: {
     keyrings: [
       {